Initial commit
As it stands I have implemented FIRST, and have some poor-mans version
of FOLLOW.  The FOLLOW implementation was based on a random example in
the tutorial I'm following
(https://serokell.io/blog/how-to-implement-lr1-parser).  This example
turned out to not actually be FOLLOW, but something else.  Now have to
fix that.