Update with alternate interpretation of FOLLOW
Rather than just recording what I see in the grammar rule for this particular
token, record transitive closure for this itemset (i.e. if Term is followed by
"*" in one rule, and terminates an 'Add' rule somewhere else, then both add the
'*' and add the entirety of FOLLOW(Add) for this itemset).