Add arguments and data structure building
As it stands I'm not doing much different because I'm using most of the same
terminal/non-terminal names.

However there is a difference in 'name' and 'int' where I can now include the
actual algebraic name and the actual integer for this argument.

This change does make the distinction between categories of inputs (e.g.
'name') and the input themselves.  It also seems to highlight the difference
between the terminals and non-terminals (where the reduction step builds a
non-terminal input from some terminals, and the shift step builds a base data
structure from a single terminal).