Add calculation of reduce rules
This seems like all the rules I needed have been determined!
N.b. I had to include something special for EOF.  For a while I way toying with a special class, until I remembered that what we're branching on in this code is not the text the user gave us -- it is the name we have given to the "type" of that text.
The tokenizer already identifies text like '123' as 'int', while still allowing the text 'int' in the user program to mean something else.
Hence I just chose the descriptive name 'EOF'.