Tie everything together
Now we generate the LR tables from a given grammar, use those tables with the tokeniser, and parse some text with that tokeniser and table combination.

Had to implement `accept' in the table generation since I'd missed this before.
Doing this made it clear that the "first" table was really quite special -- have to treat is special in yet another place.

Also reverted to using `$' for the EOF token since I'd used that everywhere else.