Adjusted the Minus sign implementation to be nullable
This was the goal I originally had -- to use Minus sign as a nullable symbol in order to test out that the nullable symbol handling worked.
noticed that as long as I put the Minus just before the terminals in the Term definition I was no longer making a left-recursive grammar.
Now things work nicely -- except for the fact that I still generate empty 'Minus' parses (which I still suspect is a feature of having nullable symbols.

Don't prefer the generation I get from this, but it exercises more code so I'm going with it.

N.b. since I'm removing the question mark thing I'm also removing the question mark from the tokenised.