Little refactoring for my integration tests
Problem here is that I've not refactored for a long while, initial approach was thinking about things as a script, now things are turning into modules to re-use in different ways and I've not refactored in between.
This is a small improvement, at least makes me able to write some integration tests.
|Using `manual_tables.py' almost as a class instead of a module -- initialising it by loading it in `generator_take2.py' and calling an initialisation function in it, then using it elsewhere.
This is not great -- I should really take the global state out of the `manual_tables.py' functions and use them without needing to adjust stuff, but this is less work for the moment.