Looks like implemented LR(1) parsing
Have not tested this thoroughly, so highly likely to be some problems left to fix, but this at least appears to be doing what it should.
As it turns out the changes were mostly mechanical for everything except finding a FOLLOW set for each itemset.
Determining the follow set for a particular itemset is quite different to determining the follow set as a whole.
Some predictions in the itemset do not get updated etc.