Only update the follow set of symbols which are "next" in a prediction
Do this because these are the only predictions that we added.
If we found out that a given prediction was nullable, and found information about the non-terminal after it, then that would be stored on a prediction that is not actually corresponding to expanding this symbol second in a prediction.
This doesn't have an affect with the current grammars, but I suspect it could cause some bugs in others.