Fix bug in `accept`
When writing this function in the first place it seems I assumed that there is only one expression when accepting the text at the end.
With this new test grammar I've defined a `Start` symbol that has more than one expression in its expansion.
This meant I wasn't getting the full expression printed out at the end.
Here we accept "the current tree" rather than "the top expression on the current tree".
That means we print out the entire tree we have parsed so far.