Add a tokeniser implementation that can take a specification for exactly what to treat as tokens
This should make it much easier to try my parser on different grammars.
Do hope to add something that takes a specification of a grammar and extracts the relevant tokens, I believe I'll have to adjust the syntax of the grammar files I'm using.
Once that's done I hope to be able to parse directly from a grammar specification.