Update test case output after commit 3df9c25e9bdb0
That commit fixed a bug where we would only return the last expression reduced when accepting.
That was fine when there was only one expression to reduce at the end, but when the `Start` rule consisted of multiple expressions together that meant we would not return the full `Start` expression, but just the last expression of that rule.
I apparently did not run the tests after having made that change -- have just noticed that this change broke some tests, and am updating the "accepted" tests now.